refactor: 중간 동작 테스트 , MYSQL 오타수정 ,ERD 수정사항 엔티티 매핑 수정

+++ src/main/java/shop/chaekmate/core/member/entity/Grade.java
@@ -33,7 +33,7 @@ public class Grade extends BaseEntity {
     private String name;
 
     @Column(nullable = false)
-    private int pointRate;
+    private Byte pointRate;
 
     @Column(nullable = false)
     private int upgradeStandardAmount; // 승급 기준 금액
